Move desktop service contracts to shared
