+++ src/electron/main/ipc/register-desktop-ipc.ts
@@ -5,10 +5,10 @@ import {
   getDesktopEventIpcChannel,
   getDesktopRequestIpcChannel,
 } from '../../../../shared/desktop-ipc'
+import type { DesktopServiceRuntime } from '../../../../shared/desktop-service-contracts'
 import { resolveConfiguredDevServerUrl } from '../../../../shared/dev-server'
 import { isTrustedRendererUrl } from '../app/navigation-security'
 import { getRendererDistDirectory } from '../runtime/app-paths'
-import type { DesktopServiceRuntime } from '../runtime/desktop-runtime-contracts'
 import type { AppUpdater } from '../updater/app-updater'
 import { createAppUpdateHandlers } from './request-handlers/app-update'
 import { createPiPackagesHandlers } from './request-handlers/pi-packages'
