+++ src/electron/main/ipc/request-handlers/pi-packages.ts
@@ -1,5 +1,5 @@
 import type { DesktopRequestHandlerMap } from '../../../../../shared/desktop-ipc'
-import type { PiThreadsService } from '../../runtime/desktop-runtime-contracts'
+import type { PiThreadsService } from '../../../../../shared/desktop-service-contracts'
 
 type PiPackagesRequestHandlers = Pick<
   DesktopRequestHandlerMap,
